+++ example/i2c_sht31_example.dart
@@ -1,0 +1,25 @@
+// Copyright (c) 2021, the Dart project authors.  Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+import 'package:dart_periphery/dart_periphery.dart';
+
+/// Grove - Temp&Humi Sensor(SHT31) is a highly reliable, accurate,
+/// quick response and integrated temperature & humidity sensor.
+void main() {
+  // Select the right I2C bus number /dev/i2c-?
+  // 1 for Raspbery Pi, 0 for NanoPi (Armbian), 2 Banana Pi (Armbian)
+  var i2c = I2C(1);
+  try {
+    var sht31 = SHT31(i2c);
+    print(sht31.getStatus());
+    print('Serial number ${sht31.getSerialNumber()}');
+    print('Sensor heater active: ${sht31.isHeaterOn()}');
+
+    var r = sht31.getValues();
+    print('SHT31 [t°] ${r.temperature.toStringAsFixed(2)}');
+    print('SHT31 [%°] ${r.humidity.toStringAsFixed(2)}');
+  } finally {
+    i2c.dispose();
+  }
+}